refactor(Jest support): Separate jest support from the rest of the library.
Separate jest support from the rest of the library. In jest.ts only install jest support, make the
tsarch library available in the index.

BREAKING CHANGE: Imports have to be changed from import {TSArch} from 'tsarch/dist/jest' to import
{TSArch} from 'tsarch'.
