+++ index.ts
@@ -2,3 +2,4 @@ export * from "./src/core/Project"
 export * from "./src/core/Result"
 export * from "./src/core/Rule"
 export * from "./src/core/TSArch"
+export * from "./src/jest/ArchMatchers"
