+++ jest.ts
@@ -1,5 +1,2 @@
-export * from "./src/core/Project"
-export * from "./src/core/Result"
-export * from "./src/core/Rule"
-export * from "./src/core/TSArch"
-export * from "./src/jest/ArchMatchers"
+import { extendJestMatchers } from "./src/jest/ArchMatchers"
+extendJestMatchers()
