+++ src/jest/ArchMatchers.ts
@@ -49,9 +49,11 @@ declare global {
 	}
 }
 
-expect.extend({
-	toPass(project?: Project, ruleToMatch?: Checkable) {
-		return toMatchArchRuleLogic(this, project, ruleToMatch)
-	}
-} as any)
+export function extendJestMatchers() {
+	expect.extend({
+		toPass(project?: Project, ruleToMatch?: Checkable) {
+			return toMatchArchRuleLogic(this, project, ruleToMatch)
+		}
+	} as any)
+}
